feat(TREDS-65)!: Refactor mantine to use theme and vars from themeToVars
