+++ src/components/ThemeProvider/ThemeProvider.tsx
@@ -1,20 +1,12 @@
-import { MantineProvider, createTheme, type MantineProviderProps } from '@mantine/core';
-import '@mantine/core/styles.layer.css';
-import '../../theme/theme.css.ts';
+// If need to remove all Mantine styling, use HeadlessMantineProvider
+import { MantineProvider, type MantineProviderProps } from '@mantine/core';
+import { theme } from '../../theme';
 
-import '@fontsource/montserrat/500.css';
-import '@fontsource/montserrat/600.css';
-import '@fontsource/montserrat/700.css';
-import '@fontsource/montserrat/800.css';
-import '@fontsource/montserrat/900.css';
-import '@fontsource/open-sans/400.css';
-import '@fontsource/open-sans/500.css';
-import '@fontsource/open-sans/600.css';
-import '@fontsource/open-sans/700.css';
-
+/*
 const theme = createTheme({
   fontFamily: 'Open Sans, sans-serif',
 });
+*/
 
 export interface ThemeProviderProps extends MantineProviderProps {}
