+++ src/theme/index.ts
@@ -1,2 +1,4 @@
-export { vars } from './theme.css.ts';
+export { vars } from './theme-old.css.ts';
+//export { vars } from './theme.css.ts';
 export { breakpoints, getComponents, themeVariables, type ThemeVariables } from './themeVariables';
+export { theme } from './theme.ts';
