+++ src/theme/theme.ts
@@ -1,0 +1,13 @@
+import { createTheme } from '@mantine/core';
+
+// Import mantine core styles
+import '@mantine/core/styles.layer.css';
+
+// Import any global CSS styles
+import '../styles/index.css';
+
+// theme
+export const theme = createTheme({
+  //fontFamily: 'Open Sans Variable, sans-serif',
+  //headings: { fontFamily: 'Montserrat Variable, sans-serif' },
+});
